refactor: simplify page components for initiatives, people, and teams
- Removed unnecessary user authentication and permission checks from the main page components for Initiatives, People, and Teams, streamlining the rendering process.
- Updated the view redirection logic to ensure default views are set to 'dashboard' for initiatives and people, and 'list' for teams.
- Enhanced dropdown components to reflect updated paths for dashboard views, improving navigation consistency across the application.
