+++ src/components/teams/teams-view-dropdown.tsx
@@ -3,11 +3,18 @@
 import { Link } from '@/components/ui/link'
 import { Button } from '@/components/ui/button'
 import { ActionDropdown } from '@/components/common/action-dropdown'
-import { Workflow, ChevronDown, Eye } from 'lucide-react'
+import { Workflow, ChevronDown, Eye, List } from 'lucide-react'
 import { useUserSettings } from '@/lib/hooks/use-user-settings'
+import { usePathname } from 'next/navigation'
+import { cn } from '@/lib/utils'
 
 export function TeamsViewDropdown() {
+  const pathname = usePathname()
   const { updateSetting } = useUserSettings()
+
+  const isList = pathname === '/teams/list'
+  const isChart = pathname === '/teams/chart'
+
   return (
     <ActionDropdown
       trigger={({ toggle }) => (
@@ -26,8 +33,25 @@ export function TeamsViewDropdown() {
       {({ close }) => (
         <div className='py-1'>
           <Link
+            href='/teams/list'
+            className={cn(
+              'flex items-center gap-3 px-3 py-2 text-sm hover:bg-accent hover:text-accent-foreground transition-colors',
+              isList && 'bg-accent'
+            )}
+            onClick={() => {
+              updateSetting('teamsDefaultView', 'list')
+              close()
+            }}
+          >
+            <List className='w-4 h-4' />
+            List
+          </Link>
+          <Link
             href='/teams/chart'
-            className='flex items-center gap-3 px-3 py-2 text-sm hover:bg-accent hover:text-accent-foreground transition-colors'
+            className={cn(
+              'flex items-center gap-3 px-3 py-2 text-sm hover:bg-accent hover:text-accent-foreground transition-colors',
+              isChart && 'bg-accent'
+            )}
             onClick={() => {
               updateSetting('teamsDefaultView', 'chart')
               close()
